+++ CHANGELOG.md
@@ -69,6 +69,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - [#332](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/332) e2e: e2e test for unbonding
 - [#331](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/331) refactor: introduce `MigrationTester`
 - [#330](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/330) fix: remove unbonding_tx_sig from UnbondedBtcDelegation
+- [#329](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/329) refactor(babylon): consolidate state and msg module
 - [#328](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/328) chore: remove slashed from BtcDelegation
 - [#326](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/326) feat: contract migration support
 - [#327](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/327) chore: remove JAIL_FOREVER